requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.3
psycopg2-binary==2.9.9
